Day 10 Complete
Fixed bug in code. Turned int datatype on numbers to float.

+++ day_10/project/calculator_app.py
@@ -33,7 +33,7 @@ operations = {
 }
 
 def calculator():
-    num1 = int(input("What's the first number?: "))
+    num1 = float(input("What's the first number?: "))
 
     # Print out operation symbols
     for symbol in operations:
@@ -43,7 +43,7 @@ def calculator():
 
     while continue_calculating:
         operation_symbol = input("Pick an operation: ")
-        num2 = int(input("What's the next number?: "))
+        num2 = float(input("What's the next number?: "))
         calc_function = operations[operation_symbol]
         answer = calc_function(num1, num2)
